created getUser Function, refactoring
