+++ src/helpers/getPreparedData.tsx
@@ -1,12 +1,12 @@
 import { Todo, TodoWithUser, User } from '../Types';
+import { getUser } from './getUser';
 
 export const getPreparedData = (
   todosFromServer: Todo[],
   usersFromServer: User[],
 ): TodoWithUser[] => {
   return todosFromServer.map(todo => {
-    const user = usersFromServer
-      .find(currentUser => currentUser.id === todo.userId);
+    const user = getUser(todo.userId, usersFromServer);
 
     return {
       ...todo,
