+++ src/helpers/getUser.ts
@@ -1,0 +1,7 @@
+import { User } from '../Types';
+
+export const getUser = (
+  id: number, users: User[],
+) => (
+  users.find(currentUser => currentUser.id === id)
+);
